HT2 renamed types/index.ts and added type to middlewares/log.ts
