+++ src/middlewares/log.ts
@@ -1,5 +1,8 @@
-// todo: google which type a function should implement
-export function log(req, res, next) {
-    console.log(`🌳🪓 ${new Date().toISOString()} 🎯 '${req.url}'`);
+import { RequestHandler } from 'express';
+
+export const log: RequestHandler = (req, res, next) => {
+    console.log(
+        `🌳🪓 ${new Date().toISOString()} 🎯 ${req.method} ${req.url}`
+    );
     next();
-}
+};
